+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.6.2
 New features:
 - Updated wasapi library
+- Add FivePointPeq biquad combo
+- Support wav with extended header
 
 Bugfixes:
 - Stop properly after failing to start with bad wasapi config