+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.6.2
+New features:
+- Updated wasapi library
+- Add FivePointPeq biquad combo
+- Support wav with extended header
+
+Bugfixes:
+- Stop properly after failing to start with bad wasapi config
+
 ## 0.6.1
 New features:
 - Add lists of supported device types in help message